pyproject.toml
[tool.poetry]
name = "reflexion-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
black = "^24.8.0"
isort = "^5.13.2"
langchain = "^0.2.12"
langchain-openai = "^0.1.20"
langgraph = "^0.2.1"
langchain-community = "^0.2.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
